+++ Runtime/IDeucarianAdjustableCameraNavigationControls.cs
@@ -1,0 +1,8 @@
+namespace Deucarian.CameraNavigation
+{
+    public interface IDeucarianAdjustableCameraNavigationControls :
+        IDeucarianCameraNavigationControls
+    {
+        float GlobalSensitivity { get; set; }
+    }
+}